requirements.txt
# Gemini Parallel Question Solver - Dependencies

# Core
google-genai>=1.0.0              # New Gemini API client (replaces deprecated google-generativeai)
Pillow>=10.0.0                # Image processing
python-dotenv>=1.0.0          # Environment variables

# CLI & UX
rich>=13.0.0                  # Beautiful terminal output
tqdm>=4.66.0                  # Progress bars

# Async utilities
aiofiles>=23.0.0              # Async file operations

# Web Server
fastapi>=0.109.0              # Modern async web framework
uvicorn>=0.27.0               # ASGI server
python-multipart>=0.0.6       # File upload support
